refactor(loss_hub): extract prepare hooks into prepare.py
Separate DiT forward input preparation from loss formulas so Flow-GRPO and NFT prepare paths live in one module while losses.py stays formula-only.
