+++ miles/backends/fsdp_utils/loss_hub/__init__.py
@@ -8,12 +8,8 @@ Swap pieces via:
 
 from miles.backends.fsdp_utils.loss_hub.advantages import grpo_normalize_rewards
 from miles.backends.fsdp_utils.loss_hub.context import DiffusionLossContext, PreparedBatch
-from miles.backends.fsdp_utils.loss_hub.losses import (
-    flow_grpo_loss_formula,
-    prepare_flow_grpo_batch,
-    resolve_loss_formula_fn,
-    resolve_prepare_fn,
-)
+from miles.backends.fsdp_utils.loss_hub.losses import flow_grpo_loss_formula, resolve_loss_formula_fn
+from miles.backends.fsdp_utils.loss_hub.prepare import prepare_flow_grpo_batch, prepare_nft_batch, resolve_prepare_fn
 
 __all__ = [
     "DiffusionLossContext",
@@ -21,6 +17,7 @@ __all__ = [
     "flow_grpo_loss_formula",
     "grpo_normalize_rewards",
     "prepare_flow_grpo_batch",
+    "prepare_nft_batch",
     "resolve_loss_formula_fn",
     "resolve_prepare_fn",
 ]
